it's work except version control

+++ common.h
@@ -7,6 +7,6 @@
 #include <stdio.h>
 #include <stdint.h>
 
-uint32_t get_bits(uint8_t start, uint8_t size, uint8_t *data);
+uint64_t get_bits(uint8_t start, uint8_t size, uint8_t *data);
 
 #endif
